WIP: Purge queue and wait

+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/services/AwsSqsClient.java
@@ -1,18 +1,26 @@
 package uk.gov.di.authentication.frontendapi.services;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.sqs.SqsClient;
 import software.amazon.awssdk.services.sqs.SqsClientBuilder;
+import software.amazon.awssdk.services.sqs.model.GetQueueAttributesRequest;
 import software.amazon.awssdk.services.sqs.model.PurgeQueueRequest;
+import software.amazon.awssdk.services.sqs.model.QueueAttributeName;
 import software.amazon.awssdk.services.sqs.model.SendMessageRequest;
 
 import java.net.URI;
 import java.util.Optional;
 
+import static java.lang.Thread.sleep;
+
 public class AwsSqsClient {
+
+    private static final Logger LOG = LoggerFactory.getLogger(AwsSqsClient.class);
 
     private final SqsClient client;
     private final String queueUrl;
@@ -40,7 +48,24 @@ public class AwsSqsClient {
     }
 
     public void purge() {
-        PurgeQueueRequest request = PurgeQueueRequest.builder().queueUrl(queueUrl).build();
-        client.purgeQueue(request);
+        GetQueueAttributesRequest attributesRequest = GetQueueAttributesRequest
+                .builder()
+                .queueUrl(queueUrl)
+                .attributeNames(QueueAttributeName.APPROXIMATE_NUMBER_OF_MESSAGES)
+                .build();
+        var result = client.getQueueAttributes(attributesRequest);
+        result.getValueForField(QueueAttributeName.APPROXIMATE_NUMBER_OF_MESSAGES.name(), Integer.class)
+                .ifPresent( count -> {
+                    LOG.info("Found {} messages in the queue", count);
+                    if (count > 0) {
+                        PurgeQueueRequest request = PurgeQueueRequest.builder().queueUrl(queueUrl).build();
+                        client.purgeQueue(request);
+                    }
+                    try {
+                        sleep(60000);
+                    } catch (InterruptedException ignored) {
+
+                    }
+                });
     }
 }
